@@ -1,18 +1,22 @@
-﻿using MichelMichels.DobissSharp.Models;
+﻿using MichelMichels.DobissSharp.Api.Models;
 using System.Diagnostics.CodeAnalysis;
 
-namespace MichelMichels.DobissSharp.Comparers
+namespace MichelMichels.DobissSharp.Comparers;
+
+internal class SubjectComparer : IEqualityComparer<Subject>
 {
-    internal class SubjectComparer : IEqualityComparer<Subject>
+    public bool Equals(Subject? x, Subject? y)
     {
-        public bool Equals(Subject x, Subject y)
+        if (x is null || y is null)
         {
-            return x.Channel == y.Channel && x.Address == y.Address;
+            return false;
         }
 
-        public int GetHashCode([DisallowNull] Subject obj)
-        {
-            return obj.Address.GetHashCode() + obj.Channel.GetHashCode();
-        }
+        return x.Channel == y.Channel && x.Address == y.Address;
+    }
+
+    public int GetHashCode([DisallowNull] Subject obj)
+    {
+        return obj.Address.GetHashCode() + obj.Channel.GetHashCode();
     }
 }
